refactor: 게임 실행과 관련된 로직은 GameController클래스로 분리
